+++ CMakeLists.txt
@@ -2,33 +2,30 @@ cmake_minimum_required(VERSION 3.10)
 # based on https://cliutils.gitlab.io/modern-cmake/
 
 project(SimMap VERSION 0.1
-        DESCRIPTION "Example project by Jens Klimke"
+        DESCRIPTION "A traffic simulation map framework"
         LANGUAGES C CXX)
 
 set(CMAKE_CXX_STANDARD 14)
 
 
 # unset testing for libs
-set(ODRLIB_ENABLE_TESTING OFF CACHE BOOL "Unset testing for odrparser")
-
-# add libraries
-include_directories(lib/Eigen)
-add_subdirectory(lib/odrparser)
-include_directories(lib/odrparser/include/)
+option(BUILD_TESTS "Sets or unsets the option to generate the test target" OFF)
 
 # set include dir
-set(SIMMAP_INCL_DIR ${PROJECT_SOURCE_DIR}/include)
+set(simmap_INCLUDE_DIR ${PROJECT_SOURCE_DIR}/include)
+set(simmap_SOURCE_DIR ${PROJECT_SOURCE_DIR}/src)
+set(Eigen_INCLUDE_DIR ${PROJECT_SOURCE_DIR}/lib/Eigen)
+set(odrparser_INCLUDE_DIR ${PROJECT_SOURCE_DIR}/lib/odrparser)
 
+# add odrparser
+add_subdirectory(lib/odrparser)
+
+# add source folder
 add_subdirectory(src)
 
 
-# first check if tests is enabled
-if(NOT DEFINED SIMMAP_ENABLE_TESTING)
-    set(SIMMAP_ENABLE_TESTING true)
-endif(NOT DEFINED SIMMAP_ENABLE_TESTING)
-
 ### GTEST
-if(SIMMAP_ENABLE_TESTING)
+if(BUILD_TESTS)
 
     # add gtest
     add_subdirectory(lib/gtest)
