+++ src/CMakeLists.txt
@@ -4,3 +4,6 @@ add_subdirectory(graph)
 add_subdirectory(server)
 add_subdirectory(odradapter)
 add_subdirectory(SimMap)
+
+include_directories(${simmap_INCLUDE_DIR})
+include_directories(${Eigen_INCLUDE_DIR})
