+++ src/base/poly1.h
