+++ src/curve/C3Spline.h
