+++ src/curve/Curve.h
