+++ src/curve/GeoElement.h
