+++ src/curve/Poly3.h
