+++ src/curve/Spiral.h
