+++ src/odradapter/CMakeLists.txt
@@ -1,7 +1,7 @@
 
 # get all header files
 file(GLOB HEADER_FILES
-        ${SIMMAP_INCL_DIR}/odradapter/*.h
+        ${simmap_INCLUDE_DIR}/odradapter/*.h
         )
 
 set(SOURCE_FILES
@@ -18,5 +18,5 @@ set(SOURCE_FILES
         _parseRoadMarking.cpp)
 
 add_library(odradapter STATIC ${SOURCE_FILES})
-target_include_directories(odradapter PRIVATE ${SIMMAP_INCL_DIR} ${SIMMAP_INCL_DIR}/odradapter)
+target_include_directories(odradapter PRIVATE ${simmap_INCLUDE_DIR} ${simmap_INCLUDE_DIR}/odradapter)
 target_link_libraries(odradapter PRIVATE odr server curve)
