+++ src/server/MapCoordinate.h
