+++ tests/base/CMakeLists.txt
@@ -1,6 +1,6 @@
 # get all header files (only done to register the headers to the project)
 file(GLOB HEADER_FILES
-            ${SIMMAP_INCL_DIR}/base/*.h
+            ${simmap_INCLUDE_DIR}/base/*.h
         )
 
 # source files
@@ -13,7 +13,7 @@ set(SOURCE_FILES
 
 # build test executable
 add_executable(BaseTest ${SOURCE_FILES} ${HEADER_FILES})
-target_include_directories(BaseTest PRIVATE ${SIMMAP_INCL_DIR})
+target_include_directories(BaseTest PRIVATE ${simmap_SOURCE_DIR} ${Eigen_INCLUDE_DIR})
 
 # add tests
 target_link_libraries(BaseTest PUBLIC gtest gmock gtest_main)
